List size incorrect after delete middle node
In qtest cmd mode, dm command do_dm() in qtest.c
for executing q_delete_mid() to delete middle node of list.

But there is no subtract 1 in size of list after dm command executed,
then, showing error message below:

"ERROR: Computed queue size as %d, but correct value is %d"

the "correct value" amid this message
is not match to the real size of the list shows on the moniter,

this error message is caused by "lcnt", "cnt" not match in do_size(),
so, by adding a line of code in do_dm() to solve this problem.