Provide an option to show/hide Shannon entropy
The concept of information entropy was introduced by Claude Shannon in
his 1948 paper "A Mathematical Theory of Communication", and is also
referred to as Shannon entropy. Shannon's theory defines a data
communication system composed of three elements: a source of data, a
communication channel, and a receiver. The "fundamental problem of
communication" – as expressed by Shannon – is for the receiver to be
able to identify what data was generated by the source, based on the
signal it receives through the channel.

In qtest, execute "option entropy 1" before command "show" which will
display both the value of each element and its Shannon entropy.

For the sake of performance, the integer-only calculation is used for
the kernel of Shannon entropy.